Reading F90 namelist data now always requires a factory
 - Enforces the principle that (in a given context) there is a
one-to-one mapping between namelist group names and Java types that can
as a minimum read and write the data, and that declare a static Java
variable for the name of each namelist group variable. Done by removing
the read(File) method from the public interface, and requiring a mapping
to be specified together with the file containing namelist format data
in a constructor.
 - Also tided up and updated some Javadoc
 - See SDCP-74